+++ gauss_cuda/cuda/cuda_gauss.h
@@ -5,6 +5,6 @@
 #include <driver_types.h>
 
 void substractLines(double *res, double *pivot, double *resVector,
-        double *pivotVector, size_t coef, size_t size, cudaStream_t stream);
+        double *pivotVector, double coef, size_t size, cudaStream_t stream);
 
 #endif
